powerpc: configs: Set initrd_high environment varible
In commit 20e072f, it force to check "bootm_low" and "bootm_size" if
"initrd_high" is missing. It will cause Linux can't boot up with ramdisk
size > 100MB on PPC.

On PPC the initrd does not need to be loaded within the boot mapping.
This assumption of what the absence of initrd_high means in commit
20e072f is not universally appropriate.

As the discussion is continuing in the community, this patch is just a
workaround to fix the issue.

Signed-off-by: Alison Wang <alison.wang@nxp.com>
Signed-off-by: Prabhakar Kushwaha <prabhakar.kushwaha@nxp.com>