baremetal: Fix spi flash driver bug
The commitId of 'ca1ac16da09' add a function 'clean_bar',
this function set 'bank_sel' to 0 after reads/writes/erases memory > 16 MiB.
It will result in an error while operating memory > 16 MiB, the same
would cause the 'test_qspi' to fail.

The error occurs because bank_sel is set to 0 and bank_curr is not set to 0.
Therefore, the 'flash->bank_curr = 0;' is added to the 'clean_bar' function
to clean up the problem.

Signed-off-by: Xianyu Wei <xianyu.wei@nxp.com>