armv8: ls1028a: change edma streamid to match hw limitation
On these chips the EDMA StreamID cannot be set to an arbitrary value
as it has bits 4 and 6 always tied to 1 so redefine the StreamID value
to match this limitation. Note that StreamID 0x28 is from the DPAA2
range of StreamIDs and since there's no DPAA2 on this chip we can
safely use the value.

Signed-off-by: Laurentiu Tudor <laurentiu.tudor@nxp.com>